Prefix / Mongoose

+++ utils/message-utils.js
@@ -1,5 +1,8 @@
 function deleteMessage(message) {
+    if (!message) return;
+
     setTimeout(() => {
+        if (message.deleted) return; // Already deleted
         message.delete();
     }, process.env.TIMEOUT);
 }
